pyproject.toml
[project]
name = "quark"
version = "0.1.0"
description = "Vanilla language modeling playground"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.5.0",
    "hydra-core>=1.3.2",
    "matplotlib>=3.10.8",
    "pandas>=3.0.0",
    "python-dotenv>=1.2.1",
    "torch>=2.10.0",
    "transformers>=5.1.0",
    "wandb>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.15.1",
    "pre-commit>=4.5.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["models"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["C", "E", "F", "I", "W", "D", "N"]
ignore = ["E741", "W605", "C408", "E402", "F722", "N806", "N803", "D107"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
docstring-code-format = true
